Update terminalmessaging.css
I love that you were able to match the text messaging nature of it. But some messages from the controller are probably better to be left justified, and the forced width of 400px doesn't scale well, especially because I have the octoprint wide screen plugin enabled. I'm dialing it back a bit here, and I think it has a lot of the functionality, while missing all the flair, but it is useful for things like M122 from the TMC drivers or M503 with all the machine settings.